audio: Store available devices in global context
